access get_cat_num_inheritance_setting from queryfieldspec apply_special_filter_cases
